Renamed hostmetrics metric metadata files from "constants" to "metadata" and added comments
