+++ receiver/hostmetricsreceiver/internal/scraper/networkscraper/network_metadata.go
@@ -25,10 +25,14 @@ const (
 	stateLabelName     = "state"
 )
 
+// direction label values
+
 const (
 	receiveDirectionLabelValue  = "receive"
 	transmitDirectionLabelValue = "transmit"
 )
+
+// descriptors
 
 var metricNetworkPacketsDescriptor = createMetricNetworkPacketsDescriptor()
